Use the Zipkin URLConnectionReporter from new zipkin-reporter libs rather than copy-pasting from brave
